Renaming Metadb to SMILE
- refactored package names
- refactored class names and references to 'metadb' to 'smile'
