+++ src/test/java/org/mskcc/smile/config/MockDataConfig.java
@@ -1,4 +1,4 @@
-package org.mskcc.cmo.metadb.config;
+package org.mskcc.smile.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.BufferedReader;
@@ -11,10 +11,10 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import org.mockito.Mockito;
-import org.mskcc.cmo.common.FileUtil;
-import org.mskcc.cmo.metadb.model.MockJsonTestData;
-import org.mskcc.cmo.metadb.service.impl.ValidRequestCheckerImpl;
-import org.mskcc.cmo.metadb.service.util.RequestStatusLogger;
+import org.mskcc.smile.commons.FileUtil;
+import org.mskcc.smile.model.MockJsonTestData;
+import org.mskcc.smile.service.impl.ValidRequestCheckerImpl;
+import org.mskcc.smile.service.util.RequestStatusLogger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
@@ -22,7 +22,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.core.io.ClassPathResource;
 
 @Configuration
-@ComponentScan(basePackages = "org.mskcc.cmo.common.*")
+@ComponentScan(basePackages = "org.mskcc.smile.commons.*")
 public class MockDataConfig {
     private final ObjectMapper mapper = new ObjectMapper();
     private final String MOCKED_REQUEST_DATA_DETAILS_FILEPATH = "data/mocked_request_data_details.txt";
